+++ lib/mix/lib/releases/models/release.ex
@@ -26,6 +26,7 @@ defmodule Mix.Releases.Release do
       include_erts: true,
       include_src: false,
       include_system_libs: true,
+      included_configs: [],
       strip_debug_info: false,
       plugins: [],
       overlay_vars: [],